Make backup-restore open source
