+++ graphql/admin/endpoints_ee.go
@@ -13,223 +13,6 @@
 package admin
 
 const adminTypes = `
-	input BackupInput {
-
-		"""
-		Destination for the backup: e.g. Minio or S3 bucket.
-		"""
-		destination: String!
-
-		"""
-		Access key credential for the destination.
-		"""
-		accessKey: String
-
-		"""
-		Secret key credential for the destination.
-		"""
-		secretKey: String
-
-		"""
-		AWS session token, if required.
-		"""
-		sessionToken: String
-
-		"""
-		Set to true to allow backing up to S3 or Minio bucket that requires no credentials.
-		"""
-		anonymous: Boolean
-
-		"""
-		Force a full backup instead of an incremental backup.
-		"""
-		forceFull: Boolean
-	}
-
-	type BackupPayload {
-		response: Response
-		taskId: String
-	}
-
-	input RestoreInput {
-
-		"""
-		Destination for the backup: e.g. Minio or S3 bucket.
-		"""
-		location: String!
-
-		"""
-		Backup ID of the backup series to restore. This ID is included in the manifest.json file.
-		If missing, it defaults to the latest series.
-		"""
-		backupId: String
-
-		"""
-		Number of the backup within the backup series to be restored. Backups with a greater value
-		will be ignored. If the value is zero or missing, the entire series will be restored.
-		"""
-		backupNum: Int
-
-		"""
-		All the backups with num >= incrementalFrom will be restored.
-		"""
-		incrementalFrom: Int
-
-		"""
-		If isPartial is set to true then the cluster will be kept in draining mode after
-		restore. This makes sure that the db is not corrupted by any mutations or tablet moves in
-		between two restores.
-		"""
-		isPartial: Boolean
-
-		"""
-		Path to the key file needed to decrypt the backup. This file should be accessible
-		by all alphas in the group. The backup will be written using the encryption key
-		with which the cluster was started, which might be different than this key.
-		"""
-		encryptionKeyFile: String
-
-		"""
-		Vault server address where the key is stored. This server must be accessible
-		by all alphas in the group. Default "http://localhost:8200".
-		"""
-		vaultAddr: String
-
-		"""
-		Path to the Vault RoleID file.
-		"""
-		vaultRoleIDFile: String
-
-		"""
-		Path to the Vault SecretID file.
-		"""
-		vaultSecretIDFile: String
-
-		"""
-		Vault kv store path where the key lives. Default "secret/data/dgraph".
-		"""
-		vaultPath: String
-
-		"""
-		Vault kv store field whose value is the key. Default "enc_key".
-		"""
-		vaultField: String
-
-		"""
-		Vault kv store field's format. Must be "base64" or "raw". Default "base64".
-		"""
-		vaultFormat: String
-
-		"""
-		Access key credential for the destination.
-		"""
-		accessKey: String
-
-		"""
-		Secret key credential for the destination.
-		"""
-		secretKey: String
-
-		"""
-		AWS session token, if required.
-		"""
-		sessionToken: String
-
-		"""
-		Set to true to allow backing up to S3 or Minio bucket that requires no credentials.
-		"""
-		anonymous: Boolean
-	}
-
-	type RestorePayload {
-		"""
-		A short string indicating whether the restore operation was successfully scheduled.
-		"""
-		code: String
-
-		"""
-		Includes the error message if the operation failed.
-		"""
-		message: String
-	}
-
-	input ListBackupsInput {
-		"""
-		Destination for the backup: e.g. Minio or S3 bucket.
-		"""
-		location: String!
-
-		"""
-		Access key credential for the destination.
-		"""
-		accessKey: String
-
-		"""
-		Secret key credential for the destination.
-		"""
-		secretKey: String
-
-		"""
-		AWS session token, if required.
-		"""
-		sessionToken: String
-
-		"""
-		Whether the destination doesn't require credentials (e.g. S3 public bucket).
-		"""
-		anonymous: Boolean
-
-	}
-
-	type BackupGroup {
-		"""
-		The ID of the cluster group.
-		"""
-		groupId: UInt64
-
-		"""
-		List of predicates assigned to the group.
-		"""
-		predicates: [String]
-	}
-
-	type Manifest {
-		"""
-		Unique ID for the backup series.
-		"""
-		backupId: String
-
-		"""
-		Number of this backup within the backup series. The full backup always has a value of one.
-		"""
-		backupNum: UInt64
-
-		"""
-		Whether this backup was encrypted.
-		"""
-		encrypted: Boolean
-
-		"""
-		List of groups and the predicates they store in this backup.
-		"""
-		groups: [BackupGroup]
-
-		"""
-		Path to the manifest file.
-		"""
-		path: String
-
-		"""
-		The timestamp at which this backup was taken. The next incremental backup will
-		start from this timestamp.
-		"""
-		since: UInt64
-
-		"""
-		The type of backup, either full or incremental.
-		"""
-		type: String
-	}
 
 	type LoginResponse {
 
@@ -460,16 +243,6 @@ const adminTypes = `
 
 const adminMutations = `
 
-	"""
-	Start a binary backup.  See : https://dgraph.io/docs/enterprise-features/#binary-backups
-	"""
-	backup(input: BackupInput!) : BackupPayload
-
-	"""
-	Start restoring a binary backup.  See :
-		https://dgraph.io/docs/enterprise-features/#binary-backups
-	"""
-	restore(input: RestoreInput!) : RestorePayload
 
 	"""
 	Login to Dgraph.  Successful login results in a JWT that can be used in future requests.
@@ -542,8 +315,4 @@ const adminQueries = `
 	queryUser(filter: UserFilter, order: UserOrder, first: Int, offset: Int): [User]
 	queryGroup(filter: GroupFilter, order: GroupOrder, first: Int, offset: Int): [Group]
 
-	"""
-	Get the information about the backups at a given location.
-	"""
-	listBackups(input: ListBackupsInput!) : [Manifest]
 	`
